done: jwt implementing & publishing

+++ src/main/java/org/baopen753/bookingappbackend/services/serviceservice/ServiceService.java
@@ -1,11 +1,13 @@
 package org.baopen753.bookingappbackend.services.serviceservice;
-
 import org.baopen753.bookingappbackend.entities.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ServiceService {
-    Service updateServicePrice(Integer serviceId, double servicePrice);
+    Service updateServicePrice(Integer serviceId, BigDecimal servicePrice);
+
     Service getServiceById(Integer serviceId);
+
     List<Service> getAllServices();
 }
